+++ src/backend/demo/tests/test_commands_create_demo.py
@@ -28,7 +28,9 @@ def test_commands_create_demo():
     """The create_demo management command should create objects as expected."""
     call_command("create_demo")
 
-    assert models.User.objects.count() == TEST_NB_OBJECTS["users"]
+    assert (
+        models.User.objects.count() == TEST_NB_OBJECTS["users"] + 1
+    )  # Monique Test (quick fix for e2e)
     assert models.Team.objects.count() == TEST_NB_OBJECTS["teams"]
     assert models.TeamAccess.objects.count() >= TEST_NB_OBJECTS["teams"]
     assert mailbox_models.MailDomain.objects.count() == TEST_NB_OBJECTS["domains"]