+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- 🔥(teams) remove search users by trigram
+
 ### Fixed
 
 - 💚(ci) improve E2E tests #492